Outsource kube_auth()
